Finish refactoring of the complex number function unit tests
